+++ jenkins-client/src/main/java/com/offbytwo/jenkins/client/util/EncodingUtils.java
@@ -9,11 +9,10 @@ public final class EncodingUtils {
 
     public static String encode(String pathPart) {
         // jenkins doesn't like the + for space, use %20 instead
-        String escape = UrlEscapers.urlPathSegmentEscaper().escape(pathPart);
-        return escape;
+        return UrlEscapers.urlPathSegmentEscaper().escape(pathPart);
     }
 
-    public static String encodeParam(String pathPart) {
+    public static String formParameter(String pathPart) {
         // jenkins doesn't like the + for space, use %20 instead
         return UrlEscapers.urlFormParameterEscaper().escape(pathPart);
     }