+++ jenkins-client/src/test/java/com/offbytwo/jenkins/client/util/EncodingUtilsTest.java
@@ -1,6 +1,5 @@
 package com.offbytwo.jenkins.client.util;
 
-import org.assertj.core.api.Assertions;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -32,25 +31,25 @@ public class EncodingUtilsTest {
 
     @Test
     public void encodeParamShouldReturnEncodedExclamationMarkDoubleQuoteAmpersampSpace() {
-        String result = EncodingUtils.encodeParam("!\"& ");
+        String result = EncodingUtils.formParameter("!\"& ");
         assertThat(result).isEqualTo("%21%22%26+");
     }
 
     @Test
     public void encodeParamShouldReturnNotEncodeSafeChars() {
-        String result = EncodingUtils.encodeParam("-_.*");
+        String result = EncodingUtils.formParameter("-_.*");
         assertThat(result).isEqualTo("-_.*");
     }
 
     @Test
     public void encodeParamShouldReturnEncodedUmlautAndOthers() {
-        String result = EncodingUtils.encodeParam("äöü#{}");
+        String result = EncodingUtils.formParameter("äöü#{}");
         assertThat(result).isEqualTo("%C3%A4%C3%B6%C3%BC%23%7B%7D");
     }
 
     @Test
     public void encodeParamShouldReturnEncodedCharacters() {
-        String result = EncodingUtils.encodeParam("-._~!$'()*,;&=@:+");
+        String result = EncodingUtils.formParameter("-._~!$'()*,;&=@:+");
         assertThat(result).isEqualTo("-._%7E%21%24%27%28%29*%2C%3B%26%3D%40%3A%2B");
     }